Rename the public types file
