+++ src/components/scenes/PasswordLoginScene.tsx
@@ -47,7 +47,7 @@ import { ChallengeModal } from '../modals/ChallengeModal'
 import { GradientFadeOut } from '../modals/GradientFadeout'
 import { QrCodeModal } from '../modals/QrCodeModal'
 import { TextInputModal } from '../modals/TextInputModal'
-import { CreateAccountType } from '../publicApi/types'
+import { CreateAccountType } from '../publicApi/publicTypes'
 import { Airship, showError } from '../services/AirshipInstance'
 import { Theme, useTheme } from '../services/ThemeContext'
 import { FilledTextInput, FilledTextInputRef } from '../themed/FilledTextInput'
